requirements.txt
psycopg2-binary==2.9.1
requests==2.27.1
twilio==7.8.0
djangorestframework==3.13.1
django-environ==0.8.1
dj-database-url==0.5.0